235 inserting binary data into blob column in sqlite fails with unhandled exception unrecognized token (#252)
* wip

* wip

* fix sqlite query #235

* sqlite create table

* sqlite alter table

* wip postgre insert binary

* postgres query

* postgres schema builder

* wip test query

* sqlite test_query

* wip postgres test query

* fix update placeholder

* postgres update delete

* postgres test query

* fix sqlite test query

* fix postgres

* wip mariadb

* wip mariadb

* wip

* wip mariadb binary insert

* wip test.c

* success mysql query

* mariadb test

* aggregate test

* rm test_query

* wip mysql test query

* mysql test query

* delete not used import

* wip test surreal query

* wip surrealdb get test success

* surrealdb insert test success

* surrealdb update test success

* surrealdb delete test success

* surrealdb raw test success

* wip surreal aggregate test

* success surreal query test

* delete unused surreal import

* success surreal test

* sqlite schema test

* postgres schema test

* mariadb schema test

* wip mysql schema

* mysql schema test

* fix test run code

* fix rm binary code

* postgres transaction test

* sqlite transaction test

* mariadb transaction test

* mysql transaction test

* rm unused code

* postgres return type test

* sqlite return type test

* mariadb return type test

* mysql return type test

* fix surreal autoincrement

* remove unused

* delete old backups

* compile for only uses database

* delete unnecessary isExists

* fix warning

* nimble docs

* insert binary test

* fix

* fix test

+++ allographer.nimble
@@ -20,9 +20,9 @@ when (NimMajor, NimMinor) > (1, 6):
 import strformat, os
 
 task test, "run testament test":
-  exec "testament p 'tests/test_*.nim'"
+  exec "testament p 'tests/*/test_*.nim'"
   for kind, path in walkDir(getCurrentDir() / "tests"):
-    if not path.contains("."):
+    if not path.contains(".") and path.fileExists():
       exec "rm -f " & path
 
 task docs, "Generate API documents":
